refactor param handler
